added dryrun per Charles' suggestion

+++ bin/remove_deleted_samples_from_s3.py
@@ -2,7 +2,7 @@
 import os
 import boto3
 
-ENV = 'staging'
+ENV = 'staging' # change to prod as necessary
 DB_SAMPLES_FILENAME = '/tmp/{env}_db_sample_ids'.format(env=ENV)
 S3_SAMPLES_FILENAME = '/tmp/{env}_s3_sample_paths'.format(env=ENV)
 
@@ -25,7 +25,7 @@ def db_sample_dict():
     return_dict[line] = 1
   return return_dict
 
-def delete_unknown_samples():
+def delete_unknown_samples(dryrun):
   s3 = boto3.resource('s3')
   pd = db_sample_dict()
   bucket='idseq-samples-{env}'.format(env=ENV)
@@ -37,16 +37,18 @@ def delete_unknown_samples():
     sample_id = line.split('/')[1]
     if sample_id not in pd:
       subpath = line.strip()
-      print('Deleting data under: {subpath}'.format(subpath=subpath))
-      resp = s3.meta.client.list_object_versions(
-          Bucket=bucket,
-          Prefix='samples/{subpath}'.format(subpath=subpath))
-      objects_to_delete = resp.get('Versions', [])
-      if len(objects_to_delete) > 0:
-        delete_keys = {'Objects' : []}
-        for item in objects_to_delete:
-          delete_keys['Objects'].append({'Key': item['Key'], 'VersionId': item['VersionId']})
-        s3.meta.client.delete_objects(Bucket=bucket, Delete=delete_keys)
+      print('{verb}: idseq-samples-{env}/samples/{subpath}'.format(
+        verb = 'Found' if dryrun else 'Deleting', env=ENV, subpath=subpath))
+      if not dryrun:
+        resp = s3.meta.client.list_object_versions(
+            Bucket=bucket,
+            Prefix='samples/{subpath}'.format(subpath=subpath))
+        objects_to_delete = resp.get('Versions', [])
+        if len(objects_to_delete) > 0:
+          delete_keys = {'Objects' : []}
+          for item in objects_to_delete:
+            delete_keys['Objects'].append({'Key': item['Key'], 'VersionId': item['VersionId']})
+          s3.meta.client.delete_objects(Bucket=bucket, Delete=delete_keys)
    
 if __name__ == "__main__":
   msg = 'This utility will delete all s3 sample data without db entries in the {env} environment. Type "y" or "yes" to continue.\n'.format(env=ENV)
@@ -56,4 +58,6 @@ if __name__ == "__main__":
     quit()
   fetch_sample_ids()
   fetch_s3_paths()
-  delete_unknown_samples()
+  delete_unknown_samples(dryrun=True)
+  input('Press enter to continue')
+  delete_unknown_samples(dryrun=False)
